nba-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
nba_api==1.4.1
python-dotenv==1.0.1
cachetools==5.3.3
